requirements.txt
aiohttp==3.7.4.post0
anyio==3.1.0
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
backcall==0.2.0
bleach==3.3.0
Brotli==1.0.9
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.0
Flask-Compress==1.9.0
future==0.18.2
hyperlink==21.0.0
idna==3.1
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.7.0
jupyter-server-proxy==3.0.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pi==0.1.2
pickleshare==0.7.5
plotly==4.14.3
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
retrying==1.3.3
Send2Trash==1.5.0
simpervisor==0.4
six==1.16.0
sniffio==1.2.0
terminado==0.10.0
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
txaio==21.2.1
typing-extensions==3.10.0.0
vpython==7.6.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
yarl==1.6.3
tqdm